+++ tests/dummy/app/controllers/backspace.ts
@@ -130,7 +130,6 @@ export default class BackspaceController extends Controller {
   }
 
   @tracked plugins: PluginConfig = [
-    // disabled until https://binnenland.atlassian.net/browse/GN-4147 is fixed
     firefoxCursorFix(),
     lastKeyPressedPlugin,
     tablePlugin,